solutions/quickback(Flask+Mongo)/src/python/requirements.txt
flask
flask-cors
cryptography
pyjwt
requests
gunicorn
pymongo